requirements.txt
cffi==1.14.3
joblib==0.17.0
numpy==1.19.2
pandas==1.1.3
pkg-resources==0.0.0
praat-parselmouth==0.3.3
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.21.3
scipy==1.5.2
six==1.15.0
sklearn==0.0
SoundFile==0.10.3.post1
threadpoolctl==2.1.0
